LIBSEARCH-1. Added index.html.erb file to override placeholder in theme
A decision was made to make the "index.html.erb" file in the
quick_search-umd_theme gem a placeholder, because a file that includes
references to the searchers really belongs in the same repository as the
application which defines the searchers that are included.

Added index.html.erb file to the "searchumd" application to show
actual search results.

https://issues.umd.edu/browse/LIBSEARCH-1