LIBSEARCH-43. Updated library dependency and removed unnecessary lines
When building the image from the Dockerfile, the build process was
failing, with a message indicating that the "libmysqlclient-dev"
package could not be found.

Changed the package to "default-libmysqlclient-dev". Not sure why there
was a change, but believe it might be caused by something in the
upstream image.

Reformatted some of the lines, and removed the addition of the
".ssh" subdirectory in "app" home, because it was not necessary.

https://issues.umd.edu/browse/LIBSEARCH-43