LIBSEARCH-53. Added Apache Nutch/Solr configuration files for JSON-LD
Added Nutch configuration to use the "umd-parsefilter-jsonld" plugin
from the UMD Apache Nutch implementation.

Modified the Solr schema.xml by adding a "documentContent" field.

Modified Nutch Solr indexing configuration so that the "documentContent"
field placed in the metadata by the "umd-parsefilter-jsonld" plugin is
placed in the "documentContent" field in Solr.

https://issues.umd.edu/browse/LIBSEARCH-53