LIBSEARCH-42. Moved website "native" interface into searchumd
Moved the controller and views for the website native interface out
of the quick_search-library_website_searcher into searchumd.

This simplifies the searcher, and avoids having to maintain separate
copies of the view templates.

Modified routes.rb, removed QuickSearchLibraryWebsiteSearcher engine,
because it was no longer needed, and added additional routes to satisfy
requirements in the view layouts (when used with the "native" interface)
to have access to certain routes.

https://issues.umd.edu/browse/LIBSEARCH-42