LIBSEARCH-15 Modified "no_results_link" configuration for searchers
Added "no_results_link" properties to each of the searcher configuration
YAML files in the config/searchers/ directory. This enables the
"no_results_link" property to be parameterized using environment
variables, if necessary.

In the en.yml file, replaced "no_results_link" entries with nothing
so that the properties in the searcher configuration YAML files will
be used.

https://issues.umd.edu/browse/LIBSEARCH-15