LIBSEARCH-119. Varnish ESI Proof-of-concept - Part 2
Added backend to enable Varnish to find the wwwdev.lib.umd.edu server.
This was somewhat suggested in

https://varnish-cache.org/lists/pipermail/varnish-misc/2016-May/024954.html

None of the documentation seems to indicate that this is necessary,
but it seems to be, as otherwise the ESI just goes against the
localhost server (search.local).

WARNING: This proof-of-concept does not actually work correctly,
because Varnish is not able to perform ESI on "https" URLs.

https://issues.umd.edu/browse/LIBSEARCH-119