LIBSEARCH-119. Nginx, SSI, proxy_pass Proof-of-concept
A proof-of-concept for using Nginx, SSI, and proxy_pass to
include content from another server into a searchumd web page.

Note: In this proof-of-concept, the "_footer.html.erb" overrides the
same file from the "quick_search-umd_theme" gem. In the actual
implementation, the change would be made in the "quick_search-umd_theme"
gem itself.

https://issues.umd.edu/browse/LIBSEARCH-119