LIBSEARCH-143. Library website native interface link now uses "query"
Modified the library website native interface link to use "query"
instead of "q" for specifying the search term.

This makes the native interface link work with Hippo, where "query"
HTTP parameter is used for the search term.

Modified the website controller in searchumd to also expect the
"query" parameter, instead of "q".

https://issues.umd.edu/browse/LIBSEARCH-143