LIBSEARCH-147. Updated WorldCat config to point to better no results URL
Update the "no_results_link" for the world_cat_discovery_api and
world_cat_discovery_api_article searchers to point at a URL that does
not display a warning about an empty query term.

Both searchers still point at the same URL, because there does not
appear to be a way to set a URL that can be limited to articles, for
the "article" searcher.

https://issues.umd.edu/browse/LIBSEARCH-147