LIBSEARCH-150. Moved searchers into searchumd
Moved the following searchers into searchumd:

From "searchumd-special":

* quick_search-archives_space_searcher
* quick_search-drum_searcher
* quick_search-internet_archive_searcher
* quick_search-fedora_searcher
* quick_search-maryland_map_searcher

from searchumd "eds-develop" branch:

* quick_search-ebsco_discovery_service_api_searcher

Added the searcher config files for the added searchers.

Modified the searcher page to show all the searchers.

Re-ordered searcher configuration in "env_example" so searchers would be
in alphabetical order.

Removed en.yml customizations, except where needed to display the
correct text. This was done so that on the searcher results page, it was
clearer which searcher was returning which results.

https://issues.umd.edu/browse/LIBSEARCH-150