LIBSEARCH-168. Add "umd_crawl" script to perform Nutch crawl
Added a "umd_crawl" script to allow multiple steps to be performed
when doing the crawl, instead of relying solely on the "bin/crawl"
script the ships with Apache Nutch.

This is needed so that we can seed the crawl database from the
sitemap.xml file, and then crawl it.

https://issues.umd.edu/browse/LIBSEARCH-168