LIBSEARCH-183. Updated gems to correct security vulnerabilites.
Updated gems to correct vulnerabilities listed by "bundler-audit" gem:

```
> bundle update rails
> bundle update puma
> bundle update sanitize
> bundle update kaminari
> bundle update bibtex-ruby
```

There is one remaining vulnerability, CVE-2020-8130, related to the
version of the "rake" gem. It does not seem feasible at this point to
update the application for this CVE.

The optimal solution is to update the "rake" gem to v12.3.3, or later.

Unfortunately, a number of dependencies have the "rake" gem "pinned" to
v10, so Bundler does not want to update the "rake" gem.

Since there are no Rake tasks in the "searchumd" application, and the
"quick_search" upstream application does not appear to use Rake for
parsing user-provided files, it seems unlikely that this vulnerability
will affect the application.

https://issues.umd.edu/browse/LIBSEARCH-183