LIBSEARCH-193. Added "JVM_MAX_RAM_PERCENTAGE" to Nutch script
Added a "JVM_MAX_RAM_PERCENTAGE" variable to the Nutch script, which
defaults to 90.0, if not otherwise set. Modified the start script to
supply the "JVM_MAX_RAM_PERCENTAGE" variable to the -XX:MaxRAMPercentage
Java system property, so that the JVM will use more than 25% of the
available RAM in the container.

https://issues.umd.edu/browse/LIBSEARCH-193