LIBSEARCH-206. SSL Root CA certification fix
On October 1, 2021, a root CA certificate that the "http_client" relied
on for verifying SSL certificates expired. The "http_client" ships with
its own set of "trusted" root CA certificates which apparently haven't
been updated for a number of years.

The expired certificate was preventing a number of searchers from making
HTTP requests.

As at least a short-term fixed, followed the suggestion in
"http_client" issue 445 to "monkey-patch" the HTTPClient class to
use the CA certificates from the system.

https://issues.umd.edu/browse/LIBSEARCH-206