LIBSEARCH-205. Fixed searchumd GUI when port forwarding
Replaced "/search" in "RAILS_RELATIVE_URL_ROOT" environment variable
with "/", because the "/search" subpath is no longer necessary.
The "/search" subpath was used when searchumd was its own front-end
application, in order allow a search home page to be provided. Since
searchumd is no longer a front-end application, but is used purely
as a backend, the "/search" subpath is not needed.

Also removed the "SCRIPT_NAME" environment variable, because it does
not appear to be used, or necessary.

https://issues.umd.edu/browse/LIBSEARCH-205